refactor: JobPosting repository 병합
